Switch default settings to central module
puzzlord.settings looks a lot like a settings file, and is in fact the
default settings file. However, projects will generally set
DJANGO_SETTINGS_MODULE to settings.dev or the like.
puzzlord.settings is really only used as a regular Python module
to provide the SITE_PASSWORD (and should be deleted to move
SITE_PASSWORD to Django settings again).